Generify EllipsoidFeatureTest by introducing a new class AbstractEllipsoidFeatureTest
